Fix travel and autoexplore pushing boulders around
There seems to be two sources for this bug:

1. The dotravel command was setting its uim level via
   `exploration_interaction_status` which, according to the comment in
   `hack.c` is meant mainly for farmove and autoexplore.

2. travel pathing was taking us through boulders intentionally, even
   though once the uim set correctly.

The solution is to avoid boulders entirely when our occupation is
traveling and to set the uim properly to nointeraction.

Note that this does not 'fix' Shift movement pushing the bolder but
leaving you a square less than you should be and it still allows
autoexplore to climb onto boulders. It also appears to allow Ctrl
movement to climb onto boulders or squeeze past boulders. What if any of
that behavior is normal or desirable I'm not so sure but this at least
prevents boulders from being pushed by travel which, IMO, is the most
critical bug, especially in regards to sokoban.

Functional Tests
----------------

- [x] wizard mode to sokoban, travel through any boulders. Note that
  you're stopped.
- [x] Ctrl-move towards a boulder. Note that you're stopped.
- [x] travel directly onto a boulder. Note that you're stopped.