Fix Run Boulder Displacement Bug
This removes the check for whether the `run` command had been
interrupted, which caused us to return from `domove` prematurely and
muck with our `x` and `y` directions.

I don't understand why the check was there in the first place, although
I'm assuming the `lookaround` is important for fixing the issue where we
weren't turning corners. I did not test without the `lookaround` call.

Functional Tests
----------------

- [x] Running around unknown corridors.

      Behaves like I expect to at least.

- [x] Run into an an unseen boulder in a corridor.

      The boulder is pushed and `@` moves onto the square previously
      occupied by the boulder.

- [x] Run into a known boulder in a room or corridor.

      The boulder is pushed and `@` moves onto the square previously
      occupied by the boulder.